Add support for lookup-refs
Changed the way that matches are found, but this is definitely more
inefficient than it needs to be. Probably no longer need the `:subs`/
`:reverse-subs` split anymore either, because we could just check for
the reverse lookups as we build the `matchers` we need to use.

Need to collapse the implementations of the forward/reverse matchers a
bit, and start building them up in a less wasteful way.